(#28) "delambdafy" pass
* Define the AstPass helper phase.

* Implements the DelambdafyRule.

Rules are compiler phases that must always run, as they are necessary for the correctness of subsequent phases.

* Implement the majority of the AST dispatch in AstPass.h.

* Extend the number of AST nodes with a replace() implementation.

* Fill out the rest of AstPass.

* Hacky subexpression parsing fix.

(expr) was interpreted as an empty tuple.

* First delambdafy test.

* Fixups.

* Introduce the isUpdated flag in AstPass.

If a pass raises the flag, the pass will re-start from the beginning. This is to prevent individual runs from interfering with each other (e.g. modifying compound statements during iteration).

* Test fixup

* Extends the delambdafy test.

* Adds another delambdafy test, and a test for an operator precedence bug.